Fix all compilation errors exposed after test exclusion
Now that the main project correctly excludes test files, we can see
the actual compilation errors in the source code. These errors have
existed all along but were masked by the xUnit errors.

FIXES APPLIED:

1. MessageBox Ambiguity (2 files)
   - src/App.xaml.cs: Added using MessageBox = System.Windows.MessageBox;
   - src/UI/SettingsWindow.xaml.cs: Added using MessageBox = System.Windows.MessageBox;
   - Resolves: CS0104 - ambiguous reference between WPF and WinForms

2. Missing System.IO References (2 files)
   - src/UI/NotificationManager.cs: Added using System.IO;
   - src/Core/ConversionQueue.cs: Added using System.IO;
   - Resolves: CS0103 - Path/File does not exist in current context

3. Type Conversion Issue (1 file)
   - src/Core/ConversionEngine.cs: Removed (uint) cast
   - Changed: image.Quality = (uint)_settings.JpegQuality;
   - To: image.Quality = _settings.JpegQuality;
   - Resolves: CS0266 - Cannot implicitly convert uint to int

All errors found by enhanced code analysis (/p:EnableNETAnalyzers=true)
These are legitimate bugs that should have been caught earlier.

CONFIDENCE: 95% - These are standard C# compilation errors with
straightforward fixes. Build should succeed now.

+++ src/Core/ConversionEngine.cs
@@ -59,7 +59,7 @@ public class ConversionEngine
 
                 using var image = new MagickImage(sourcePath);
                 image.Format = MagickFormat.Jpeg;
-                image.Quality = (uint)_settings.JpegQuality;
+                image.Quality = _settings.JpegQuality;
 
                 // Preserve EXIF orientation
                 image.AutoOrient();
